requirements.txt
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
boto3==1.35.29
botocore==1.35.29
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.10.0
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
eval_type_backport==0.2.0
executing==2.1.0
fastapi==0.115.0
fastavro==1.9.7
fastjsonschema==2.20.0
filelock==3.16.1
firebase==4.0.1
firebase-admin==6.5.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gcloud==0.18.3
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.66.0
grpcio-status==1.66.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-oauth==0.15.1
httpx-sse==0.4.0
huggingface-hub==0.25.1
idna==3.7
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
kaleido==0.2.1
kiwisolver==1.4.7
lida==0.0.14
llama-cloud==0.1.4
llama-index==0.11.22
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.22
llama-index-embeddings-adapter==0.2.2
llama-index-embeddings-openai==0.2.5
llama-index-experimental==0.4.0
llama-index-finetuning==0.2.1
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.2.1
llama-index-llms-mistralai==0.2.5
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-postprocessor-cohere-rerank==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.13
llmx==0.0.21a0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
mdurl==0.1.2
mistralai==1.1.0
mistune==3.0.2
mizani==0.11.4
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
openai==1.50.2
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
patsy==0.5.6
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
plotnine==0.13.6
portalocker==2.10.1
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.0rc1
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyogrio==0.10.0
pyparsing==3.1.2
pypdf==4.3.1
pyproj==3.6.1
Pyrebase4==4.8.0
python-dateutil==2.8.2
python-jwt==4.1.0
python-multipart==0.0.11
python-pptx==0.6.23
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.1.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.38.6
statsmodels==0.14.3
streamlit==1.36.0
striprtf==0.0.26
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.20.0
toml==0.10.2
toolz==0.12.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.45.1
typer==0.12.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.31.0
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
wordcloud==1.9.3
wrapt==1.16.0
XlsxWriter==3.2.0
yarg==0.1.9
yarl==1.9.4
